Treat LOAD_LEVEL_START the same as level re-enter.
If we are loading LOAD_LEVEL_START autosave with existing level file,
this means we have re-entered this level from another level, and we need
to run it for 10 seconds.

Fixes buggy mgu3m2 -> mgu3m3 -> mgu3m2 transition if after re-entering
mgu3m2 player is killed and level is reloaded.