Work around games closing standard handles on Windows
Some game libraries (admittedly, old ones) are linked against CRTDLL.DLL,
which seems to close the standard handles on shutdown, breaking the system
console if running a dedicated server.
Work around this by stowing away the original standard input handles and
working on duplicates, and attempt to re-duplicate the handles on certain
errors.