Allocate packet entities per client.
Size of packet entities circular buffer must be power of two to handle
entity counters wraparound correctly. With globally allocated buffer
this is only the case when ‘maxclients’ is power of two.

To make it work with any ‘maxclients’, switch to per-client packet
entities allocation. This also minimizes server memory usage if not all
client slots are in use.