Update github build/release workflows
Reuse the build.yml for building binaries. There may be some reason why
the upstream has the build script in .ci/nightly.sh but it lacks mac
builds and some others.

Here the build workflow is modified to upload the built binaries as
github artifacts and then another job which runs only on main-branch
pushes (PR merges, direct pushes to main) and uploads them to the
"latest" release.